Graphical installer (for Debian based distribution)
use:
buildDebPackage.sh <tarball>
Build the package which will includes the content of the tarball.

This first release consists in building a Debian package wrapping the
voxin-installer and launching it non-interactively (-l).

So that if needed, the user may install Voxin without typing any
command line: e.g. the user download the link via the web browser, and
once downladed the Debian based distribution should propose to install
the Debian package.