Calling a voxin module installed locally
Speech-dispatcher (>=0.11) can launch a local text-to-speech module
stored under the user's account at:
~/.local/libexec/speech-dispatcher-modules

Voxin supplies a shell-script as local module which will call the
binary module (sd_voxin) stored in the user's installation directory.

The current file is a template of this shell script; the variables
between % have just to be replaced.