voxin_install_all.sh: install all the Voxin tarballs
Install all the Voxin tarballs included in the SRC_DIR directory.
Note that these tarballs will first be unpacked in the $BUILD_DIR
directory.

If the -u option is supplied, all the Voxin voices are uninstalled.
Then optionally, the tarballs included in SRC_DIR are installed.

If this script is launched as super user, the voices will be installed
system-wide.
If launched as regular user, the voices will be installed locally.